+++ source/AskFi.Runtime.Persistence/StorageEnvironment.cs
@@ -1,3 +1,6 @@
+using AskFi.Runtime.Persistence.Encoding;
+using AskFi.Runtime.Persistence.InMemory;
+
 namespace AskFi.Runtime.Persistence;
 
 public sealed class StorageEnvironment : IStorageEnvironment
